Major data model updates
Must accommodate rooms and allow front desk to get a snapshot
(including timer state) of the entire office. This means we need a
TreatmentArea abstraction (a room, or part of a room), and timers will
be on two different objects (Treatments using Machines or whatever in
TreatmentAreas). You also have to be able to start/pause/resume. Hence
the ProcessTimer abstraction and polymorphic "process" association (so
that all the timer logic is in one place, and you can extend timers to
other objects in the future). Add labels to measurements, since you can
have multiple sets per session (e.g., "Before" and "After"). Add a
display name to the machine (for front desk). Treatment sessions have
to be associated with a Machine in order for the admin/tech/front desk
to query what treatments are ongoing in a facility.