Dashboard treatment areas
Did this break on styling? Each treatment area has to be a separate
form; otherwise it can't tell which one fired (and will always take the
first one). "Always take the first" is correct when each is its own
form, since there will only ever be one timer object in the form. Also
add Expire Timer, which the dashboard predates.